Count-occurrences-of-a-word-in string
Count occurrences of a word in string
You are given a string and a word your task is that count the number of the occurrence of the given word in the string and print the number of occurrence of the word. 

Approach:

First, we split the string by spaces in a
Then, take a variable count = 0 and in every true condition we increment the count by 1
Now run a loop at 0 to length of string and check if our string is equal to the word
if condition true then we increment the value of count by 1 and in the end we print the value of count.